add option to edit bookmark to the BookmarkForm component, correctly export form types, update bookmarks and categories hook
