+++ src/forms/BookmarkForm/index.ts
@@ -1,0 +1,2 @@
+export { BookmarkForm } from './BookmarkForm';
+export type { BookmarkValues } from './BookmarkForm';
