+++ src/forms/CategoryForm/index.ts
@@ -1,1 +1,2 @@
 export { CategoryForm } from './CategoryForm';
+export type { CategoryValues } from './CategoryForm';
