--- src/forms/NewBookmarkForm/index.ts
@@ -1,1 +1,0 @@
-export { NewBookmarkForm } from './NewBookmarkForm';
